// src/types/commission.ts
export type CommissionLevel = {
  id: 'DA220C06-F828-45CD-9D34-532937AE11B0';
  commission: 40;
  level: 0;
  created_by: 'A5EB68DD-58F1-4E3F-9449-36404A0FC65C';
  updated_by: 'A5EB68DD-58F1-4E3F-9449-36404A0FC65C';
  created_at: '2024-06-15T17:24:15.330Z';
  updated_at: '2024-06-15T17:24:15.330Z';
};